+++ lib/src/commands/create/create.dart
@@ -19,6 +19,8 @@ final _templates = [
   FlutterPkgTemplate(),
   FlutterPluginTemplate(),
   VeryGoodDartCLITemplate(),
+  VeryGoodDocsSiteTemplate(),
+  VeryGoodFlameGameTemplate(),
 ];
 
 final _defaultTemplate = _templates.first;