+++ test/src/commands/create/create_test.dart
@@ -26,12 +26,14 @@ const expectedUsage = [
       '''                              (defaults to "A Very Good Project created by Very Good CLI.")\n'''
       '''    --executable-name         Used by the dart_cli template, the CLI executable name (defaults to the project name)\n'''
       '    --org-name                The organization for this new project.\n'
-      '                              (defaults to "com.example.verygoodcore")\n'
+      '''                              (defaults to "com.example.verygoodcore")\n'''
       '''-t, --template                The template used to generate this new project.\n'''
       '\n'
       '''          [core] (default)    Generate a Very Good Flutter application.\n'''
       '''          [dart_cli]          Generate a Very Good Dart CLI application.\n'''
       '          [dart_pkg]          Generate a reusable Dart package.\n'
+      '          [docs_site]         Generate a Very Good documentation site.\n'
+      '          [flame_game]        Generate a Very Good Flame game.\n'
       '          [flutter_pkg]       Generate a reusable Flutter package.\n'
       '          [flutter_plugin]    Generate a reusable Flutter plugin.\n'
       '\n'
@@ -727,7 +729,7 @@ void main() {
 
       group('valid template names', () {
         Future<void> expectValidTemplateName({
-          required String getPackagesMsg,
+          String? progressLog,
           required String templateName,
           required MasonBundle expectedBundle,
           required String expectedLogSummary,
@@ -779,9 +781,9 @@ void main() {
             progressLogs,
             equals(['Generated ${generatedFiles.length} file(s)']),
           );
-          verify(
-            () => logger.progress(getPackagesMsg),
-          ).called(1);
+          if (progressLog != null) {
+            verify(() => logger.progress(progressLog)).called(1);
+          }
           verify(() => logger.created(expectedLogSummary)).called(1);
           verify(
             () => generator.generate(
@@ -826,7 +828,7 @@ void main() {
 
         test('core template', () async {
           await expectValidTemplateName(
-            getPackagesMsg: 'Running "flutter packages get" in .tmp/my_app',
+            progressLog: 'Running "flutter packages get" in .tmp/my_app',
             templateName: 'core',
             expectedBundle: veryGoodCoreBundle,
             expectedLogSummary: 'Created a Very Good App! 🦄',
@@ -835,7 +837,7 @@ void main() {
 
         test('dart pkg template', () async {
           await expectValidTemplateName(
-            getPackagesMsg: 'Running "flutter pub get" in .tmp/my_app',
+            progressLog: 'Running "flutter pub get" in .tmp/my_app',
             templateName: 'dart_pkg',
             expectedBundle: veryGoodDartPackageBundle,
             expectedLogSummary: 'Created a Very Good Dart Package! 🦄',
@@ -844,7 +846,7 @@ void main() {
 
         test('flutter pkg template', () async {
           await expectValidTemplateName(
-            getPackagesMsg: 'Running "flutter packages get" in .tmp/my_app',
+            progressLog: 'Running "flutter packages get" in .tmp/my_app',
             templateName: 'flutter_pkg',
             expectedBundle: veryGoodFlutterPackageBundle,
             expectedLogSummary: 'Created a Very Good Flutter Package! 🦄',
@@ -853,19 +855,36 @@ void main() {
 
         test('flutter plugin template', () async {
           await expectValidTemplateName(
-            getPackagesMsg: 'Running "flutter packages get" in .tmp/my_app',
+            progressLog: 'Running "flutter packages get" in .tmp/my_app',
             templateName: 'flutter_plugin',
             expectedBundle: veryGoodFlutterPluginBundle,
             expectedLogSummary: 'Created a Very Good Flutter Plugin! 🦄',
           );
         });
 
-        test('dart CLI template', () async {
+        test('dart cli template', () async {
           await expectValidTemplateName(
-            getPackagesMsg: 'Running "flutter pub get" in .tmp/my_app',
+            progressLog: 'Running "flutter pub get" in .tmp/my_app',
             templateName: 'dart_cli',
             expectedBundle: veryGoodDartCliBundle,
             expectedLogSummary: 'Created a Very Good Dart CLI application! 🦄',
+          );
+        });
+
+        test('docs site template', () async {
+          await expectValidTemplateName(
+            templateName: 'docs_site',
+            expectedBundle: veryGoodDartCliBundle,
+            expectedLogSummary: 'Created a Very Good documentation site! 🦄',
+          );
+        });
+
+        test('flame game template', () async {
+          await expectValidTemplateName(
+            templateName: 'flame_game',
+            expectedBundle: veryGoodFlameGameBundle,
+            expectedLogSummary:
+                'Created a Very Good Game powered by Flame! 🔥🦄',
           );
         });
       });